+++ src/pages/VerifyCertificate.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+function VerifyCertificate() {
+  return (
+    <div>
+        
+    </div>
+  )
+}
+
+export default VerifyCertificate
