+++ src/utils/addLeaf.js
@@ -1,75 +1,8 @@
 import { db } from "../../firebase";
 import { doc, getDoc, setDoc, arrayUnion } from "firebase/firestore";
 import generateHash from "./generateHash";
-
-function ensureEven(hashes) {
-  if(hashes.length % 2 !== 0) {
-      hashes.push(hashes[hashes.length - 1]);
-  }
-}
-
-
-const generateMerkleRoot = (hashArray)=>{
-  if(!hashArray || hashArray.length === 0){
-     return '';
-  } 
-
-  // if there are a odd number of hashes, we simply
-  ensureEven(hashArray);
-
-  const combinedHashes = [];
-  
-  for(let i=0; i<hashArray.length; i+=2){
-    const hashPairconcatenated = hashArray[i] + hashArray[i+1];
-    const newHash = generateHash(hashPairconcatenated);
-    combinedHashes.push(newHash);
-  }
-
-  if(combinedHashes.length === 1){
-    return combinedHashes.join('');
-  }
-
-  return generateMerkleRoot(combinedHashes);
-
-}
-
-const generateMerkleTree = async (hashes,merkleRoot) =>{
-  if(!hashes || hashes.length === 0) {
-    return [];
-  }
-  let lvl = 0;
-  const tree = {};
-
-  tree[lvl] = hashes;
-
-  const generate = (hashes, tree) => {
-    if(hashes.length === 1) {
-        return hashes;
-    }
-    ensureEven(hashes);
-    const combinedHashes = [];
-    for(let i = 0; i < hashes.length; i += 2) {
-        const hashesConcatenated = hashes[i] + hashes[i + 1];
-        const hash = generateHash(hashesConcatenated);
-        combinedHashes.push(hash);
-    }
-    lvl++;
-    tree[lvl] = combinedHashes;
-    return generate(combinedHashes, tree);
-  }
-  generate(hashes, tree);
-  const merkleTreeRef = doc(db, 'merkleTree', 'merkleTreeStructure');
-
-  try {
-    await setDoc(merkleTreeRef,{
-      merkleRoot: merkleRoot,
-      merkleTree: tree
-    },{merge: true})
-  } catch (error) {
-    console.log("Error Generating Merkle Tree",error);
-  }
-}
-
+import generateMerkleRoot from "./generateMerkleRoot";
+import generateMerkleTree from "./generateMerkleTree";
